Update the instructions to connect to Android
Later versions of Android have two extra options:

'Minimum TLS version'. This should be set to v1.2 because others are insecure
'Online Certificate Status', which uses OCSP to check if our certificate is
revoked. As we don't use  that I recommend setting this to 'Do not verify'.

Both option can be shown or not, depending on the version of Android used